refactor: change redux store structure
