+++ src/shared/redux/index.ts
@@ -1,13 +1,3 @@
-import { configureStore } from '@reduxjs/toolkit';
-import alertReducer from './slices/alertSlice';
-
-const store = configureStore({
-  reducer: {
-    alertState: alertReducer,
-  },
-});
-
-export default store;
-
-export type RootState = ReturnType<typeof store.getState>;
-export type AppDispatch = typeof store.dispatch;
+export * from './store';
+export * from './StoreProvider';
+export * from './hooks';
