+++ src/widgets/authorization/ui/Logout.tsx
@@ -1,6 +1,6 @@
 import { Button } from '@mui/material';
 import { logout } from '../model/firebase';
-import { useAppDispatch } from '@shared/hooks/hook';
+import { useAppDispatch } from '@shared/redux';
 import { showAlert } from '@shared/redux/slices/alertSlice';
 import { AlertStyle } from '@widgets/alert/model/Alert.model';
